backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
greenlet==3.1.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.20
httpx==0.28.1
openpyxl==3.1.5
reportlab==4.2.5
